Pub: Implement `findPackageManagerDependencies()`
Implement the `findPackageManagerDependencies()` function and return
that Pub must run before Gradle if any Gradle definition files seem to
be associated to a Pub Flutter project, because in this case it is
likely that the Gradle project depends on the `local.properties` file
generated by Pub.

This is a preparation for replacing the current mechanism where Pub
calls Gradle directly with a different mechanism where the Gradle
dependencies are inserted into the Pub dependency tree in a
post-processing step.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>