DatabaseUtils: Hold the created HikariDataSource in a map
This commit introduces a map as a new property, which holds the created
HikariDataSource for the given PostgresConfiguration. In this way only a new
HikariDataSource is created, when it doesn't exist for the given
PostgresConfiguration. Otherwise, the already created one from the map will be
used. This prevents creating new database connections for the same
PostgresConfiguration.

Signed-off-by: Onur Demirci <onur.demirci@bosch.io>