PubFunTest: Reduce test scope for the Android and CocoaPods test project
Only compare the Pub project and the packages it depends on, but ignore
the Android project itself. This reduces the size of the expected
results file as less packages have to be contained, and because only the
relevant scope of the Android project has to be contained. This also
allows to use the dependency tree representation instead of the
dependency graph which makes it easier to read the diff if the test
fails.

As a side-effect, this fixes an issue where IntelliJ cannot show the
diff of the test, and instead shows that the test is still running even
if it is actually finished.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>