analyzer: Add a `PackageManagerDependencyHandler`
The `PackageManagerDependencyHandler` can be used to create and resolve
dependencies between different package managers.

Package managers can add a dependency on another definition file and
scope name to their dependency tree using the
`createPackageManagerDependency()` function. This dependency is resolved
when all package managers have finished, by searching for projects from
the named definition file and adding a dependency to each of these
projects, with the transitive dependencies defined by the scope name.

This is implemented as an operation on the dependency graph, to provide
better performance than if it was implemented on the dependency tree.

The resolution of dependencies between different package managers takes
place in the `AnalyzerResultBuilder`. With this change dependency graphs
which are not reachable from any project are not contained anymore in
the result. the `AnalyzerResultBuilder` needs to be adapted to account
for that change, as one test added a dependency graph which was not
linked to any project.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>