model: Introduce a `concludedCopyrights` field to package curations
Similar to the `concludedLicense` field, introduce a `concludedCopyrights`
field that is to be set exclusively via a package curation [1] to
override any detected copyright statements. Note that there is no such
thing as *declared* copyright statements because package managers do not
support them explicitly [1].

The concluded copyrights will get associated to all effective licenses;
there currently is no way to curate a copyright statement for a specific
license only. Behavior-wise this is no change compared to previous
feature of curating authors and enabling the `addAuthorsToCopyrights`
option.

As a bonus, this fixes a subtle bug where previously packages might have
been skipped in the scan if `authors` were set but
`addAuthorsToCopyrights` was disabled.

Resolves #4519.

[1]: https://github.com/oss-review-toolkit/ort/pull/5504#issuecomment-1191168485
[1]: https://github.com/oss-review-toolkit/ort/blob/main/docs/config-file-curations-yml.md

Signed-off-by: Rainer Bieniek <extern.rainer.bieniek@porsche.de>
Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>