SpdxDocumentFile: Remove duplicate packages created for projects
Currently, under certain circumstances the SPDX package manager can
create projects and packages with the same identifiers. Prevent this
by filtering out affected packages when creating the final
PackageManagerResult.

This is a quick fix for this issue. Since there is some discussion
ongoing about the heuristic how to classify an artifact as a project
or package, this may be tackled differently in the future though.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>