Analyzer: Rewrite support for Yarn2+
The previous design, while working on simple projects, was not
compatible with complex projects relying on Yarn workspaces : The
'node_modules' linker was creating symbolic links pointing to symbolic
links, pointing back to the former, therefore walking such file
hierarchies with cycles crashed ORT.
This commit changes completely the design of the package manager. It
tries to use as much as possible Yarn wrapper commands to build the
dependency graph.
First, `yarn info` is called to get the dependency tree of the project.
While providing a lot of information, it unfortunately does not contain
package details and dependency scope information.
To alleviate the former problem, `yarn npm info` is then called with all
packages, and details are fetched from the registry.
Regarding the lack of dependency scope information, the project
'package.json' files are parsed to find out the scope of project
dependencies. Please note that only project dependencies have their
scope qualified. Dependencies of dependencies are always falling in the
'dependencies' scope.
Another shortcoming is that Yarn 2+ does not provide details of a
project : thus, `npm list` has to be run to get this information.

A functional test showcasing support of workspaces will be added in a
future commit.

Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.io>