Pub: Use the `PackageManagerDependencyHandler` for Gradle dependencies
For Flutter projects, `Pub` uses a heuristic to determine which `Gradle`
build files seem to belong to the Flutter project. Before this change,
`Pub` was executing `Gradle` to analyze these projects and insert their
dependencies into the dependency tree of the `Pub` project. Also,
`Gradle` had to filter out those files, to avoid that they are analyzed
twice. The downside of this approach was that `Gradle` had to know about
this implementation detail of `Pub`, and that if a similar feature would
have to be introduced for other package managers (like it will be for
`Pub` dependencies on `CocoaPods` in future), the implementation could
quickly become unmaintainable.

To improve this, make use of the new mechanism to resolve dependencies
on projects from other package managers introduced in the previous
commit.

Note that `Pub` will still call `Gradle` to analyze `Gradle` projects
contained in Flutter packages, because these definition files are not
part of the project repository.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>