helper-cli: Add helper command for merging analyzer result files
This helper CLI function is used by Porsche to solve the issue #4364

The rationale behind this is that some projects at Porsche deliver individual
analyzer-results for each subproject in a large monorepo. The FOSS analyst needs
to see a condensed form of the individual dependency graphs across the project
monorepo. We solve this issue by merging all individual analyzer results into one.

Signed-off-by: Rainer Bieniek <extern.rainer.bieniek@porsche.de>