Pub: Filter packages which are also projects
If multiple Pub projects depend on each other, the Pub implementation
does not know that the dependency is a project dependency, therefore
also a package with the same identifier is created. As a quick fix,
remove those packages before creating the final result.

The multi-module test project now shows an issue where the scopes of one
project are not correctly resolved if another previously analyzed
project already had a dependency on the same package. This will be
addressed in a follow-up commit.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>