Pub: Do not store the processed packages globally
Before this change, `Pub` maintained a list of all package names which
have already been added to a dependency tree as class property. This
lead to the issue that only the first occurence of each package was
shown in the dependency tree. This was especially an issue if multiple
`Pub` projects shared dependencies, because then these dependencies did
only show up in the first project that depended on them.

To fix this issue, replace the class property with a parameter of the
`buildDependencyTree()` function. This way the dependency trees are
complete, while still preventing infinite loops.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>