client/osv: Add the known ecosystem values to the model
Define the string constants for all currently known ecosystem values
according to the listing in [1].

Note that for this purpose no enum is used in order to avoid limiting
the ecosystem's value domain in the model and query interface. So, no
module / library code change and upgrade is needed when the ecosystem
value domain changes. This may happen when new data source get
integrated, the `serverUrl` gets overridden to switch to a different
endpoint like staging or to even a self-hosted instance with a custom
set of data sources.

[1] https://ossf.github.io/osv-schema/

Signed-off-by: Frank Viernau <frank_viernau@epam.com>