Analyzer: Add support for PNPM
PNPM [1] is an alternative package manager for JavaScript / TypeScript
projects.
To detect PNPM projects, the `pnpm-lock.yaml` file is being used.

By using the `--shamefully-hoist` option when installing packages using
PNPM, a directory structure similar to NPM and Yarn is created for the
`node_modules` directory. This enables ORT to reuse the NPM
implementation to resolve dependencies.

[1]: https://pnpm.io

Resolves #2024.

Signed-off-by: Marcel Bochtler <marcel.bochtler@bosch.io>