ClearlyDefinedPackageCurationProvider: Filter duplicate curations
When requesting curations for a certain coordinate, the ClearlyDefined
API [1] sometimes also returns curations for other coordinates. For
example, the response for "maven/mavencentral/org.ow2.asm/asm/7.0" also
contains curations for other versions of the same package, and also for
other packages that start with "asm", like "asm-tree". This means, that
the response may contain duplicate curations, if multiple coordinates
are sent at once.

As a result, it can happen that ORT adds the same curation multiple
times to a package. To prevent this, filter duplicate curations before
returning the result.

[1]: https://api.clearlydefined.io/api-docs/#/curations/post_curations_

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>