Osv: Somewhat improve the mapping of the severity
ORT's data model for the severity of a vulnerability is quite generic.
The wide flexibiltiy it provides makes it unclear how to best map the
obtained values from the provider to that data model.

The use of `VulnerabilityReference.severity` in [1] indicates that a
float score value is a valid option. So, change the mapping to float for
now as that is easier to understand - even though this is a lossy
transformation.

In the longer run, ORT's vulnerability data model should be improved to
support vectors in a less generic stricter typed way.

[1] https://github.com/oss-review-toolkit/ort/blob/470e2f323b516389bd96a63a81c67b74891a36f8/model/src/main/kotlin/VulnerabilityReference.kt#L58-L63

Signed-off-by: Frank Viernau <frank_viernau@epam.com>