Osv: Add a fallback for the severity
The server does return vulnerabilities which do not have a severity
value in the dedicated property. The unspecified `databaseSpecific`
property often times holds a primitive `severity` property with values
such as `[HIGH, MEDIUM, LOW]`. Make use of these values as a fallback as
these to provide more indication than a `null` value.

Note: The data model of 'osv/client' currently uses subtypes of
JsonElement to expose a couple of unspecified JSON objects as
properties. Accessing these requires the client code to add
'kotlinx.serialization' as dependency which is not nice. A solution to
that would be to use "raw" string values containing the JSON, which is
unfortunately not yet possible but may become so in the future, see
[1][2][3].

So, for now add 'kotlinx.serialization' as dependency to the advisor in
order to access the property and leave a FIXME comment as reminder.

[1] https://github.com/Kotlin/kotlinx.serialization/issues/1298
[2] https://github.com/Kotlin/kotlinx.serialization/issues/1405
[3] https://github.com/Kotlin/kotlinx.serialization/issues/1058

Signed-off-by: Frank Viernau <frank_viernau@epam.com>