Analyzer: Add Analyzer for PNPM
PNPM [1] is a package manager for node projects. It tries to save disk
space by storing dependencies in a global `node_modules` directory and
creating hard links from the global store to the project's
`node_modules` directory.
Due to this approach it is not completely compatible with ORT's NPM
Analyzer, which relies on a specific `node_modules` directory structure.

The PNPM Analyzer uses the `pnpm` CLI's `list` command to collect all
the relevant packages. This works for "simple" projects, as well as for
"workspace" projects with nested projects.

As the `pnpm-lock.yaml` file is the only way to distinguish between PNPM
and NPM projects, PNPM projects always require the lock file. Without
it, NPM will try to resolve the dependencies.

[1]: https://pnpm.io

Resolves #2024.

Signed-off-by: Marcel Bochtler <marcel.bochtler@bosch.io>