MavenDependencyHandler: Improve the checks for project dependencies
Under certain circumstances, the Maven analyzer could create a package
with the same identifier as a local project. This caused a sanity
check in AnalyzerResultBuilder to fail. To prevent this, improve the
tests whether a package corresponds to a local project in
MavenDependencyHandler. Also make sure that for the affected package
the correct linkage is returned.

Resolves #5581.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>