Revert "Add a custom deserializer for the Vulnerability class"
This reverts commit 993ba81f230ea2328a077315b584daf424be24e8.

A custom serializer makes it more expensive to work with the code.
So, drop the backwards compatibility to simplify any upcoming changes
to the vulnerability model.

Note that the backwards compatibility is for a data format dating back
to 2021 and that the value of more than a year old vulnerability data is
questionable anyway, because in the typical use case one is interrested
in up-to-date data.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>