EvaluatedModelReporterFunTest: Eliminate some redundant testing
The test input files 'evaluated-model-reporter-test-input.yml' and
'static-html-reporter-test-input.yml' are almost identical. The former
is a copy from the latter with slight adjustments on top, see [1].
So, the corresponding evaluated model reporter test cases are also
mostly redundant.

Eliminate that redundancy by undoing the introduction of the seperate
test asset, and instead using the same updated test asset file for the
multiple test cases again. This is in line with the previous approach
of using just a single test asset which covers the relevant use cases
for testing multiple components.

[1] 507bfad55868211a5a43d9263f1316d1a16d5f68

Signed-off-by: Frank Viernau <frank_viernau@epam.com>