Evaluator: Add a rule which allows examining the project's source tree
Allow access to the project's source tree in order to enable doing
arbitrary checks on files like the ones in repolinter [1]. Adding
arbitrary file checks do make the rules API more powerful, as it allows
highly customizable checks which can automate parts of the checks
typically done prior to open sourcing a project.

Not using a third-party tool makes sense as it is simpler to use,
because it sticks to a single way (rules.kts) for writing the policy rules.

Note that the implementation intentionally is limited to the project's
source tree, e.g. it does not work for dependencies, because doing
such checks throughout the dependency tree does not have obvious needs
and is not feasible anyway in terms of execution time.

[1] https://github.com/todogroup/repolinter

Signed-off-by: Frank Viernau <frank_viernau@epam.com>