example-rules: Rename the rule 'LINKED_WEAK_COPYLEFT'
The rule is nested inside a package rule, whereas only the inner rule
may flag violations. The rule name of the outer rule is unused.

Align both names for consistency with other rules and use a slightly
shorter name which does not mention all aspects of the `require`
conditions. This relaxes the need for renaming the rule in case of
adjustments to the `require` condition.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>