github/workflows: Explicitly align the runner images with Docker
Also use Ubuntu 22.04 with GitHub actions. `ubuntu-latest` currently still
utilizes Ubuntu 20.04, see [1].

This allows to upgrade to Python 3.10 and revert fa83ffc. Note that
version "3.10" must be quoted so that YAML parsing does not interpret
this as a floating point value from which the trailing 0 could be
stripped, also see [2].

[1]: https://github.blog/changelog/2022-08-09-github-actions-ubuntu-22-04-is-now-generally-available-on-github-hosted-runners/
[2]: https://github.com/actions/setup-python/issues/482

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>