Stop polluting `Any` object with a `logger` property
While using ORT's `logger` extension property on `Any` object was
convenient, it polluted also non-ORT classes and thus created the pitfall
of accidentally logging on those non-ORT classes. Though this was
circumvented by the `isOrtClass` check, it was only performed at runtime
if the respective code path was triggered. Also, that check reimplemented
the logger caching that is already part of Log4j's Kotlin API.

So remove all of ORT's custom logger wrapping in favor of using Log4j's
Kotlin API directly. This requires all classes that want to log to be
explicitly "marked" in one of the ways outlined at [1]. Choose the
second approach involving a companion object which "is generally preferable
[...] as there is a single logger created for every instance of the
class".

However, a drawback of this appraoch is that derived classes might
accidentally use the base class's logger, resulting in the wrong class
name being used in log output. While this could be avoided by making the
companion object private, that would not work with classes that already
have a (public) companion object. So, for this change it was manually
ensured that each class that uses `logger` (or `createAndLogIssue`) has
its own `Logging` companion object.

There are only two necessary exceptions to this rule:

- Kotlin `object`s cannot contain companion objects and thus need to
  implement the `Logging` interface themselves.
- To allow logging from ORT's `CliktCommand` extension functions, a
  `logger` extension property is introduced.

[1]: https://logging.apache.org/log4j/kotlin/index.html#usage

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>