Change the default configuration file from ort.conf to config.yml
Change the default format for the ORT configuration file from Hocon to
YAML. The new name "config.yml" is chosen over "ort.yml" to avoid
confusion with the repository configuration file ".ort.yml".

This is a breaking change as all users who want to continue using the
Hocon configuration file will have to pass the path to the configuration
file on the command line.

Fixes #5550.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>