Pip: Pass the detected Python version to python-inspector
Pass the detected Python major version to python-inspector. If Python 2
is the detected major version pass 2.7 because that is the only version
supported by python-inspector and also the Python 2 version installed in
the ORT Docker image. If Python 3 is the detected major version pass
3.10 because that is the version currently installed in the ORT Docker
image. This will be improved later on by using the actually installed
Python versions, but for these values are hardcoded to make sure that
the tests pass on CI.

This commit fixes the inconsistency that the python-inspector always
uses Python 3.8 for resolving the dependency tree, ignoring the version
that is used by Pip.

There could still be inconsistencies if different versions of Python are
installed. This will be improved later on by using the actually
installed Python versions, but for now the Python versions are hardcoded
to make sure that the tests pass. Also, once the python-inspector can
also provide package metadata, the locally installed Python versions
will only be required for version detection, but not for depedency
resolution anymore.

The changed dependency tree for the "Python 2 should resolve setup.py
dependencies correctly for spdx-tools-python" test shows an issue with
python-inspector which is missing some dependencies [1].

[1]: https://github.com/nexB/python-inspector/issues/62

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>