Align on using `NONE` instead of an empty SPDX package verification code
This aligns with the ScanCode result parser code. As noted in ece788a,
strictly speaking passing `SpdxConstants.NONE` as the package
verification code in not supported by the SPDX specification, but still
doing so has clearer semantics than passing an empty string. Moreover,
the package verification code will eventually anyway be removed in the
context of [1].

[1]: https://github.com/oss-review-toolkit/ort/issues/5681

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>