PathExclude: Fix matching zero segments with a leading double asterisk
Use the `FileMatcher` as it behaves as expected in that regard.

Note that as of this change the double asterisk does not match the
filename anymore, put only path segments which are directories. For
example, `**.ext` stops matching `directory/test.ext` but due to the
fix the better pattern `**/*.ext` can now be used instead. So, replace
all patterns of the from `**.ext` by a pattern of the form `**/*.ext`
in all tests and also in ORT's own `.ort.yml`.

Fixes #4357.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>