PathExcludeGenerator: Use glob expressions for matching directory names
Make the search in the project's source tree for paths to be excluded
more flexible by using globs. Any matching directory paths are then
added to the path excludes with a `/**` suffix.

Note that these globs are expanded at the runtime of the helper command
in order to create a minimal set of path excludes that actually match
paths in the project's source tree. So, while these globs could also
haven been used as path exclude patterns as-is, this is not done to
avoid having patterns that do not match anything.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>