GitWorkingTree: Properly close the repository after each operation
Previously, the `Repository` instance remained open during the lifetime
of the `GitWorkingTree`. That is a problem if some clean-up code (like
in tests) wants to delete a cloned Git repository while there is still
a GitWorkingTree instance.

To avoid that, provide a getter for the `repo` property which allows the
caller to auto-close the `Repository` instance after use as on the next
access of the `repo` property it will be re-evaluated. The added runtime
overhead due to this should be neglible compared to the duration of the
Git operation.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>