DockerFile: Introduce a new "component-based" `Dockerfile`
The main point of this new `Dockerfile` is to allow individual tools to be
installed independently in arbitrary versions, not limited to the
version that the underlying distribution comes with. These "components"
are installed in separate multi-images to benefit from Docker's parallel
building of those, and the tool binaries are then copied to the final
image. In contrast to a layered approach this avoids rebuilding if only a
single tool version changes.

Signed-off-by: Helio Chissini de Castro <helio.chissini-de-castro@bmw.de>