batect: Use a login shell to run the functional analyzer tests
Since the previous commit many tools the functional tests depend on are
configured in the user profile. Use a login shell as the entrypoint to
make sure that these can be found while running the tests.

The new entrypoint requires that the command is quoted [1], therefore it
does not work anymore to the pass additional arguments to the batect
task, because they would not be part of the quoted string. To address
this, move the arguments into the batect task and rename it to
`funTestAnalyzer` as it can now only be used to run the functional
analyzer tests.

[1]: https://batect.dev/docs/reference/config/tasks/#command

Signed-off-by: Helio Chissini de Castro <helio.chissini-de-castro@bmw.de>