scanoss: Replace the hard-coded "Blacklist" with configurable suffixes
For reference, the original hard-coded extensions from ScanOss's
deprecated Java client [1] are used.

Additionally, support for generic filename suffixes as declared as part of
the maintained Python client [2] is added, and the Python client's
extensions and suffixes are used as defaults.

This is a step towards getting rid of the dependency on the deprecated
ScanOss Java library complelety in favor of only using ORT's own client
library.

[1]: https://github.com/scanoss/scanner.java/blob/350c211/src/main/java/com/scanoss/scanner/BlacklistRules.java#L32-L40
[2]: https://github.com/scanoss/scanoss.py/blob/f27ed92/src/scanoss/scanner.py#L54-L75

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>