Git: Simplify recovery logging
`recoverCatching` also provides the original exception so there is no
need to use a separate `onFailure`. Combine those instead and log at info
level only as these are recorvery paths that do not necessarily mean
that something went wrong yet. Keep separate info log statements to
announce the fallback on a separate line with a log prefix to make it
stand out.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>