Git: Migrate fetch code to JGit
This is a step towards using JGit exclusively in order to get rid of the
Git CLI as a dependency. It also helps to avoid confusion when debugging
authentication failures as JGit and the Git CLI may use different sources
for credentials.

Note that the work-around to fetch single missing tags is not required
anymore with JGit as it behaves correctly in that regard. Similarly,
JGit does not fail when unshallowing non-shallow working trees, so the
according condition can be omitted.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>