FossIdTest: Refactor test for deleting a scan
Refactor the test that tests the deletion of a failed scan to call the
scan function separately for the two packages. This is a preparation for
changing the scan function to take only one package instead of a list.

Calling the scan functions for the packages separately works because
`FossId` stores the relevant information for deleting scans in class
properties.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>