AdvisorRecordTest: Drop vulnerabilities from a couple of test setups
The tests which are only about `issues` do setup `vulnerabilities`. This
is not necessary as `vulnerabilties` are unrelated and may create a
misleading impression that these are related.

That makes this test consistent with the test cases for vulnerabilities
which also do not create issues in their respective test setups.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>