Remove the feature to add authors to copyrights
The feature was a crude way to curate copyright statements by curating
authors and enabling the addition of authors as copyrights (with a
missing year even). That appraoch does not work if authors and copyrights
need to stay separate, as e.g. by German law not every author is a
copyright holder.

So remove this feature in favor of an upcoming change that introduces
proper and dedicated curations for concluded copyright statements.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>