Gradle: Add a task to check `NOTICE` file contents
Add a tasks that checks whether all copyright holders found in the code
base are properly captured in the `NOTICE` file, which is a prerequisite
for migrating to a generic Copyright statement. Later, the task could also
be used in CI to check whether any newly introduced Copyrighs need to be
added to the `NOTICE` file.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>