Pip: Capture metadata using python-inspector
Let `PythonInspector.run` return a `PythonInspectorResult` and add
functions to convert this result into the ORT model. This allows to
remove the code to capture package metadata from `Pip` as the metadata
is now retrieved from python-inspector.

Currently python-inspector does only provide the long descriptions of
packages [1]. Use only the first lines of those descriptions to not
bloat the ORT result, even though sometimes the first line does not
contain a meaningful description.

As python-inspector provides all available hashes for the remote
artifacts, use the strongest algorithm that is available.

python-inspector does currently not provide the package classifiers
which were previously used to get license information [2].

The previous implementation used the `.tar.gz` artifact as binary
artifact if no other artifact was available. This was wrong, because the
`.tar.gz` is the source artifact, therefore the binary artifact is now
empty in cases where only the URL of a `.tar.gz` is available.

The expected output for running the python-flask test project on Windows
is not updated in this commit, this will be addressed separately.

[1]: https://github.com/nexB/python-inspector/issues/74
[2]: https://github.com/nexB/python-inspector/issues/81

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>