WorkingTreeCache: Ignore the VCS path
Ignore the VCS path when initializing the working tree. The scanner
implementation always scans full repositories, so sparse checkout are
not required. This improves the performance if an analyzed repository
contains many projects in different directories, because previously one
working tree for each of them would have been created during scanning.
Now only one working tree is created for this repository.

This addresses the issue that if many working trees for the same
repository exist, this could lead to a lot of requests to the repository
in a short time which can cause rate limit issues in some setups.

Add a check to the `PackageProvenanceResolver` to verify that the VCS
path from the package metadata exists, because the working tree would
now not throw an exception anymore if it does not exist.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>