Pip: Do not create a virtual environment
As no dependencies are installed anymore there is no need to create a
virtual environment. The Python binary from the path is now used to get
metadata from setup.py. This might temporarily cause issues if the
setup.py is not compatible with this Python version, but this will soon
be fixed once this data can be retrieved from the python-inspector
output (this will be available once [1] is released).

[1]: https://github.com/nexB/python-inspector/pull/69

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>