Scanner: Add a fallback for resolving nested provenances
When the nested provenances cannot be resolved,
`ScanController.nestedProvenances` won't contain an entry for the
corresponding package. The code crashes as it assumes that such entries
always exist, see [1].

In case the provenance resolution of package has been successful, but
only the nested provenance resolution has failed, the crash can be
avoided by using a nested provenance containing just the root, but no
nested repositories.

Note: The mentioned assumption can still cause a crash in case resolving the
(non-nested) provenance for a package failed, which still needs to be
addressed.

[1] https://github.com/oss-review-toolkit/ort/blob/92b475be0f78933c9172caf63f9c40c1f9fd6b91/scanner/src/main/kotlin/ScanController.kt#L296

Signed-off-by: Frank Viernau <frank_viernau@epam.com>