AnalyzerRun: Drop some default values
The default values have been introduced [1] only for backwards
compatibile deserialization in 2019. As deserialization of such ORT
files is not supported anymore, this is not a breaking change.

Drop the test as it existed only for testing the removed backwards
compatibility.

Note: A similar change is planned for `EvaluatorRun`, but this requires
further refactoring, so cannot be done as part of this change.

[1] https://github.com/oss-review-toolkit/ort/commit/d1beb288b4c899babbbe08048240e50abec9bf4e

Signed-off-by: Frank Viernau <frank_viernau@epam.com>