Reporter: Slightly simplify the interface documentation
The comment mentions `OrtResult` as input, so mentioning
`AnalyzerResult` and `ScannerRecord` is a bit redundant. The previous
comment was also inconsistent as it didn't mention the stages for the
advisor and evaluator.

Don't mention the individual stages for simplifying an upcoming change.
While at it also drop the second sentence, as it is redundant.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>