model: Inline `ScanRecord`
`ScanRecord` originally existed to separate the output from the input
and for minimizing test setup. As tests have meanwhile been changed to
consistently use `ScanRecord.EMPTY`, the test setup even becomes a bit
more minimal when `ScanRecord` is inlined, as the construction of empty
`AccessStatistics` has been removed, because it became redundant.

One shortcoming of the nesting is that the code ends up with a
multitude of class names which can be easily mixed up, e.g. the classes
`ScannerRun`, `ScanRecord`, `ScanResult`, `ProvenanceScanResult` and so
forth.

From the above, it seems reasonable to eliminate one level of nesting to
simplify the data model a bit.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>