ScanController: Fix a crash when nested provenance resolution failed
The scanner crashes at [1] in case there is no entry for the given key.
This happens when the nested provenance resolution failed. Stop
returning any scan results for packages for which the nested provenance
is not available.

Note: It may make sense to return the scan results for the provenances
which are actually available, in particular the scan result
corresponding to the non-nested provenance / main code repository.

[1] https://github.com/oss-review-toolkit/ort/blob/3996c2e2a571b59044d28a93325785d6d478e948/scanner/src/main/kotlin/ScanController.kt#L296

Signed-off-by: Frank Viernau <frank_viernau@epam.com>