Bundler: Allow to configure the Bundler version to use
Use the Bundler version that comes with JRuby by default, but allow to
configure any later Bundler version to use. This avoids the need to
install Bundler at all in many cases. For the case that Bundler needs to
be installed in a specific version, skip the "already installed" check
as for that single Gem Bundler's own caching is fast enough.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>