MavenSupport: Also fix the `url` if it is an implicit value set by Maven
For POMs that do not set an SCM URL themselves, the fixup of the SCM
does not work if the parent SCM does not set an `connection`, but only
the `url`. This results in an unwanted original SCM URL.
For instance, spring-boot-starter-parent [1] only sets the `url` and not
the `connection`. This results in an original SCM `url` like
'https://github.com/spring-projects/spring-boot/childArtifactName'.

The issue this causes is, that the `Project`'s `VcsInfo` is merged with
this original SCM's URL as `fallbackUrl`, causing the path to be set to
`childArtifactName`, which is not available in the child repository.
Fix this issue, by also taking the `url` into account when replacing the
SCM with the parent's one.

[1]: https://repo1.maven.org/maven2/org/springframework/boot/spring-boot-starter-parent/2.7.4/spring-boot-starter-parent-2.7.4.pom

Signed-off-by: Marcel Bochtler <marcel.bochtler@bosch.io>