MavenSupport: Do not overwrite all child project SCM values
When maven parses the POM it uses the parent POM's SCM connection value
if the child itself doesn't set an SCM connection.
The issue in the previous implementation was that everything in the
child's SCM was replaced with the parent SCM.
So if only the SCM URL was correctly set in the child'd SCM, this value
would be overwritten by the incorrect parent SCM URL.

Fix this issue by only setting the connection, and not touching the
other values.

Signed-off-by: Marcel Bochtler <marcel.bochtler@bosch.io>